analyzer pushdown
