+++ sql/core/src/main/scala/org/apache/spark/sql/execution/debug/package.scala
@@ -29,7 +29,7 @@ import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
 import org.apache.spark.sql.catalyst.expressions.codegen.{ByteCodeStats, CodeFormatter, CodegenContext, CodeGenerator, ExprCode}
-import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, UnaryNode}
+import org.apache.spark.sql.catalyst.plans.logical.{DebugSampleColumn, LogicalPlan}
 import org.apache.spark.sql.catalyst.plans.physical.Partitioning
 import org.apache.spark.sql.catalyst.trees.TreeNodeRef
 import org.apache.spark.sql.catalyst.util.StringConcat
@@ -299,16 +299,6 @@ package object debug {
 
     override protected def withNewChildInternal(newChild: SparkPlan): DebugExec =
       copy(child = newChild)
-  }
-
-  case class DebugSampleColumn(
-    child: LogicalPlan,
-    sampleColumns: Seq[Expression]) extends UnaryNode {
-
-    override protected def withNewChildInternal(newChild: LogicalPlan): DebugSampleColumn =
-      copy(child = newChild)
-
-    override def output: Seq[Attribute] = child.output
   }
 
   case class DebugSampleColumnExec(
